Refactoring: split `budget.rs` into several files
